Only initialize Xlink library once
