+++ inbm-vision/vision-agent/vision/data_handler/request_data_handler.py
@@ -2,7 +2,7 @@
     Handles requests from data handler.
     - Validates request and checks for valid nodes
 
-    @copyright: Copyright 2021 Intel Corporation All Rights Reserved.
+    @copyright: Copyright 2021-2022 Intel Corporation All Rights Reserved.
     @license: Intel, see licenses/LICENSE for more details.
 """
