+++ inbm-vision/vision-agent/vision/tests/unit/test_status_watcher.py
@@ -48,12 +48,13 @@ class TestStatusWatcher(TestCase):
                            security=new_security,
                            status=new_status)
 
+    @patch('inbm_vision_lib.xlink.xlink_library.XLinkLibrary.__init__', return_value=None)
     @patch('inbm_vision_lib.timer.Timer.start')
     @patch('inbm_vision_lib.invoker.Invoker.__init__', return_value=None)
     @patch('vision.registry_manager.RegistryManager.__init__', return_value=None)
     @patch('vision.data_handler.data_handler.DataHandler.send_telemetry_response')
     @patch('vision.data_handler.data_handler.DataHandler.load_config_file')
-    def test_get_done_status(self, mock_load, mock_response, mock_registry, mock_invoker, mock_timer):
+    def test_get_done_status(self, mock_load, mock_response, mock_registry, mock_invoker, mock_timer, mock_xlink_lib):
         mock_data_handler = DataHandler(Mock(), Mock())
         s = StatusWatcher([self.r1, self.r2], mock_data_handler, 10)
         self.assertFalse(s.is_all_targets_done())
